+++ changelog.md
@@ -30,7 +30,7 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 
 
-[Unreleased]: https://github.com/VeryCrazyDog/mysql-parser/compare/0.1.2...HEAD
-[0.1.2]: https://github.com/VeryCrazyDog/mysql-parser/compare/0.1.1...0.1.2
-[0.1.1]: https://github.com/VeryCrazyDog/mysql-parser/compare/0.1.0...0.1.1
+[Unreleased]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.2...HEAD
+[0.1.2]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.1...0.1.2
+[0.1.1]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/VeryCrazyDog/level-logger/releases/tag/0.1.0